requirements.txt
streamlit
pandas
numpy
google-generativeai
langchain
langchain-community
langchain-google-genai
faiss-cpu
pypdfium2
pdfplumber
python-dotenv